Commented functions and moved added functions to extension.cpp extension.h
