--- testingfile.cpp
@@ -1,25 +1,0 @@
-#include <iostream>
-#include <cmath>
-#include <array>
-#include <string>
-#include <vector>
-
-int decimal_to_binary(const int& dec);
-
-int main() {
-  std::vector<std::vector<double>> ker { {1/10, 1/10, 1/10},
-                                         {1/10, 2/10, 1/10},
-                                         {1/10, 1/10, 1/10} };
-}
-
-
-
-
-
-
-
-
-
-
-
-// end
